Modif Compte Update

+++ S2-01_Developpement_Application/DailyBank/src/main/java/application/view/ComptesManagementViewController.java
@@ -102,15 +102,24 @@ public class ComptesManagementViewController {
 
 	@FXML
 	private void doModifierCompte() {
-
+		int selectedIndice = this.lvComptes.getSelectionModel().getSelectedIndex();
+		if (selectedIndice >= 0) {
+			CompteCourant cpt = this.oListCompteCourant.get(selectedIndice);
+			this.cmDialogController.editerCompte(cpt);
+		}else{
+			// Alerte théoriquement inatteignable
+			AlertUtilities.showAlert(this.containingStage,"Erreur de sélection", "Aucun compte ne semble sélectionné!","Vous devez sélectionner un compte à supprimer!!",AlertType.ERROR);
+		}
 	}
 
 	@FXML
 	private void doSupprimerCompte() {
 		int selectedIndice = this.lvComptes.getSelectionModel().getSelectedIndex();
 		if (selectedIndice >= 0) {
-			CompteCourant cpt = this.oListCompteCourant.get(selectedIndice);
-			this.cmDialogController.supprimerCompte(cpt);
+			if(AlertUtilities.confirmYesCancel(containingStage, "Suppression de compte", "Le compte ne pourra pas être retrouvé", "Voulez-vous réellement le supprimer?", AlertType.CONFIRMATION)){
+				CompteCourant cpt = this.oListCompteCourant.get(selectedIndice);
+				this.cmDialogController.supprimerCompte(cpt);
+			}
 		}else{
 			// Alerte théoriquement inatteignable
 			AlertUtilities.showAlert(this.containingStage,"Erreur de sélection", "Aucun compte ne semble sélectionné!","Vous devez sélectionner un compte à supprimer!!",AlertType.ERROR);
@@ -131,15 +140,14 @@ public class ComptesManagementViewController {
 	}
 
 	private void validateComponentState() {
-		// Non implémenté => désactivé
-		this.btnModifierCompte.setDisable(true);
 		
-
 		int selectedIndice = this.lvComptes.getSelectionModel().getSelectedIndex();
 		if (selectedIndice >= 0) {
+			this.btnModifierCompte.setDisable(false);
 			this.btnSupprCompte.setDisable(false);
 			this.btnVoirOpes.setDisable(false);
 		} else {
+			this.btnModifierCompte.setDisable(true);
 			this.btnSupprCompte.setDisable(true);
 			this.btnVoirOpes.setDisable(true);
 		}
